feat!: [#58342] remove implicit settings
BREAKING CHANGE:

Previously, users could choose not to pass a `Settings` object to
`FastRAMQPI`, solely configuring their application implicitly through
environment variables. This was never documented, as the best practice
is to specify the program's entire configuration in a single `Settings`,
with a `FastRAMQPISettings` object as a child.

The option for implicit settings has been removed, forcing users to
follow the documented best practice.