feat!: [#59332] require graphql version
BREAKING CHANGE:

Before, the GraphQL version defaulted to v3 and was set through an
environment variable. This doesn't make sense, as the application can
only be run using the version it was written against. Now, we explicitly
require a GraphQL version to be set in the code, at the same place where
the auto-generated client is passed.