feat!: [#59372] wrap lifespan in FastIntegration
This commit makes FastIntegration wrap the lifespan provided to FastAPI,
such that state yielded in the provided lifespan asynccontextmanager is
exposed to both the request state and the app state.

Why this is not the default in Starlette or FastAPI is unknown, but it
has been very confusing in the past and has lead to the need for the
context for app state.

BREAKING CHANGE:

To implement this change the initialization of the `app.state.context` on
the FastAPIIntegrationSystem had to be delayed until the lifespan has run.