fix(fastramqpi): [#61299] unify the liveness and readiness endpoints
This commit unifies the the `/health/live` and `/health/ready` endpoint
implementations, such that they both utilize the same implementation.

Before this commit the readiness probe simply returned whether the
FastAPI application was able to respond to requests, while the liveness
probe ran all FastRAMQPI healthchecks, checking backing services, etc.

Now both probes check the backing services.

The relevant documentation for how these probes are supposed to work can
be found at: https://kubernetes.io/docs/concepts/workloads/pods/pod-lifecycle/

The documentation is a bit confusing, however a few main points can be
summerized.

About the effect of the probes:

* The livenessProbe is used to decide whether the integration should be
  restarted or not.
* The readinessProbe is used to decide whether the integration should
  receive traffic or not.

About the implementation of the probes:

* The livenessProbe can be implemented however you would like.
* The readinessProbe should whether check back-end services are
  available.

Using the same probe implementation for both allows us to configure both
a readinessProbe and a livenessProbe for the integration, for instance
such that we can stop sending traffic to an integration while it has a
temporary outage, but that we will only resort to actually restarting
the integration if the situation does not seem as temporary as first
assumed.