feat(amqp): [#61280] allow declaring upstream exchanges
This commit allows AMQPSystems to connect to non-existing upstream
exchanges, declaring them if they do not exist.

This aligns the code for both exchanges, such that they
`upstream_exchange` behaves exactly as `exchange` does.

This behavior is useful if an exchange needs to both be declared and
listened to within the same integration, producing the same pattern as
for listening to external exchanges.