fix: [#61610] parse OS2mo GraphQL datetimes as Europe/Copenhagen
Even though ISO 8601 promises "unique and unambiguous" representations
of datetimes, it is impossible to unambiguously represent timestamp in a
specific time zone using only a UTC offset.

For example, the 31st October 2010 at 00:00 in Copenhagen and Cairo is
both represented as the timestamp `2010-10-31T00:00:00+02:00`. However,
one day later in Copenhagen is represented as
`2010-11-01T00:00:00+01:00` -- now with a UTC offset of 01:00 because of
daylight saving time -- whereas the timestamp in Cairo is
`2010-11-01T00:00:00+02:00`.

ISO 8601 is thus inherently a broken standard. Until RFC 9557 introduces
proper timestamps with time zone locations, the best we can do is assume
that everyone observes Copenhagen time.