feat(ariadne): [#63829] introduce ariadne-codegen plugin
This commit introduces an ariadne-codegen plugin to handle UNSET in the
generated input-types within the generated client code.

Version 0.7.1 of Ariadne codegen handles UNSET types nicely for arguments in
queries, but not within input fields, thus fields that should have been
optional and defaulting to UNSET are instead optional, but required due to
having no default value.

It is not possible to simply set the value to UNSET when constructing the
objects as Pydantic refuses to parse the UNSET variable as an optional
Employee or similar.

Future versions of Ariadne seemingly handles UNSET in input types correctly,
by utilizing the `exclude_unset` feature in Pydantic 2, however we are bound
to Pydantic 1, as MO cannot be upgraded due to FastAPI being a blocker.

Thus for now we are at an impasse that this plugin attempts to solve.

The plugin modifies the generated GraphQL client, such that input types which
are UNSET in the GraphQL schema can be either their original type or the
UnsetType while defaulting to the UNSET type.

This ensures that the fields are no longer required, and that the UNSET
behavior works as expected, as Ariadne 0.7.1 correctly strips UNSET values
from the payloads within the `execute` method of the client, as used by
arguments in queries.

See:
* https://github.com/mirumee/ariadne-codegen/issues/121
* https://github.com/mirumee/ariadne-codegen/pull/128
* https://github.com/mirumee/ariadne-codegen/issues/165
For details.

The plugin can be used by inserting it into the `tool.ariadne-codgen` section
of the `pyproject.toml` file, as shown below:
```toml
    [tool.ariadne-codegen]
    target_package_name = "autogenerated_graphql_client"
    ...
    plugins = [
        # Return values directly when only a single top field is requested
        "ariadne_codegen.contrib.shorter_results.ShorterResultsPlugin",
        # Add UNSET and UnsetType to generated input types
        "fastramqpi.ariadne.UnsetInputTypesPlugin",  # <-- This
    ]
```

This plugin can be removed once we utilize a newer version of Ariadne codegen.