feat(ramqp, mo): [#64435] add exchange parameter to MOAMQPSystem publish_message
This commit adds the `exchange` parameter to the `MOAMQPSystem`s
`publish_message` such that it resembles `_publish_message` from the
`AMQPSystem` more closely.

The change is needed to enable integrations to publish messages to a
specific queue on the `MOAMQPSystem`, which requires setting the
routing-key to the queue name and the `exchange` parameter to `""`.