feat!: [#64073] use TCP instead of ASGI app for GraphQL event handlers
The previous design was smart, but it didn't work properly. Using
sockets as our interface ensures that the interface works in the dumbest
way possible.

In the future, we should probably make the event-handling functionality
an external component.

BREAKING CHANGE:

Since we no longer bind directly to the ASGI app, uvicorn must be
configured to listen on 127.0.0.1 port 8000.